requirements.txt
# ===================================================
# 업비트 자동매매 시스템 - 필요 라이브러리 목록
# Python 3.13 호환 버전
# 설치 명령어: pip install -r requirements.txt
# ===================================================

# 업비트 API 연동
pyupbit==0.2.33

# 환경변수 (.env 파일) 읽기
python-dotenv==1.0.0

# Supabase (클라우드 DB) 연동
supabase==2.3.5

# 텔레그램 알림 봇
python-telegram-bot==20.7

# 데이터 처리 (Python 3.13 호환 버전)
pandas==2.2.3
numpy==2.2.3

# 스케줄링 (정해진 시간에 자동 실행)
schedule==1.2.1

# 로깅
loguru==0.7.2

# HTTP 요청
requests==2.31.0

# 그래프 시각화 (백테스팅 결과 차트)
matplotlib==3.9.4

# YAML 설정 파일 읽기
pyyaml==6.0.2
